Split AniListSeriesProvider.
