+++ Emby.Plugins.AniList/AniListImageProvider.cs
@@ -1,0 +1,71 @@
+using MediaBrowser.Common.Configuration;
+using MediaBrowser.Common.Net;
+using MediaBrowser.Controller.Entities;
+using MediaBrowser.Controller.Entities.TV;
+using MediaBrowser.Controller.Providers;
+using MediaBrowser.Model.Entities;
+using MediaBrowser.Model.Logging;
+using MediaBrowser.Model.Providers;
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+using MediaBrowser.Model.Serialization;
+using MediaBrowser.Model.Configuration;
+using MediaBrowser.Controller.Entities.Movies;
+
+namespace Emby.Plugins.AniList
+{
+    public class AniListImageProvider : IRemoteImageProvider
+    {
+        private readonly IHttpClient _httpClient;
+        private readonly IApplicationPaths _appPaths;
+        private readonly Api _api;
+        public AniListImageProvider(ILogManager logManager, IHttpClient httpClient, IApplicationPaths appPaths, IJsonSerializer jsonSerializer)
+        {
+            _httpClient = httpClient;
+            _appPaths = appPaths;
+            _api = new Api(logManager.GetLogger(Name), httpClient, jsonSerializer);
+        }
+
+        public string Name => "AniList";
+
+        public bool Supports(BaseItem item) => item is Movie || item is Series || item is Season;
+
+        public IEnumerable<ImageType> GetSupportedImages(BaseItem item)
+        {
+            return new[] { ImageType.Primary };
+        }
+
+        public Task<IEnumerable<RemoteImageInfo>> GetImages(BaseItem item, LibraryOptions libraryOptions, CancellationToken cancellationToken)
+        {
+            var seriesId = item.GetProviderId(ProviderNames.AniList);
+            return GetImages(seriesId, cancellationToken);
+        }
+
+        public async Task<IEnumerable<RemoteImageInfo>> GetImages(string aid, CancellationToken cancellationToken)
+        {
+            var list = new List<RemoteImageInfo>();
+
+            if (!string.IsNullOrEmpty(aid))
+            {
+                var primary = _api.Get_ImageUrl(await _api.WebRequestAPI(_api.AniList_anime_link.Replace("{0}", aid), cancellationToken));
+                list.Add(new RemoteImageInfo
+                {
+                    ProviderName = Name,
+                    Type = ImageType.Primary,
+                    Url = primary
+                });
+            }
+            return list;
+        }
+
+        public Task<HttpResponseInfo> GetImageResponse(string url, CancellationToken cancellationToken)
+        {
+            return _httpClient.GetResponse(new HttpRequestOptions
+            {
+                CancellationToken = cancellationToken,
+                Url = url
+            });
+        }
+    }
+}
