+++ Emby.Plugins.AniList/AniListMovieProvider.cs
@@ -1,0 +1,18 @@
+﻿using MediaBrowser.Common.Configuration;
+using MediaBrowser.Common.Net;
+using MediaBrowser.Controller.Providers;
+using MediaBrowser.Model.Logging;
+using MediaBrowser.Model.Serialization;
+using MediaBrowser.Controller.Entities.Movies;
+
+//API v2
+namespace Emby.Plugins.AniList
+{
+    public class AniListMovieProvider : AniListMetadataProvider<Movie, MovieInfo>
+    {
+        public AniListMovieProvider(IApplicationPaths appPaths, IHttpClient httpClient, ILogManager logManager, IJsonSerializer jsonSerializer) : base(appPaths, httpClient, logManager, jsonSerializer)
+        {
+            
+        }
+    }
+}
